signy: restructure build system to accommodate esp-idf
Restructures build system to accommodate the intricacies of how the
esp-idf component manager interprets component repositories. In order to
fetch the entire repository when using as a git dependency, the
idf_component.yml must reside at the root. We then opt to include the
CMakeLists.txt from the component directory when ESP_PLATFORM is
defined, and we move the Zephyr specific CMakeLists.txt to the module
directory. The default CMakeLists.txt path for a Zephyr module is
`zephyr/` so we opt to omit the field from the module.yml.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>